Editing the upgrade script and the doom script.
When the user upgrades doom neovim, the repo gets pulled down and placed
into /tmp/doom-neovim. All the contents are then copyed over and placed
inside of ~/.config/doom-neovim, leaving the plugins directory
untouched. Placing the repo in /tmp is a security change so if there is
a doom-neovim directory inside of the config folder it will not get
overwritten.